Update Docker Compose Gemfile.lock and setup instructions
The Gemfile lockfile didn't yet have support for ARM-based linux, so
this adds that.

The Docker Compose file had the now-deprecated `version` field, so I
removed that.

I also updated the README file to run the setup and test instructions
against the Docker container instead of outside the docker container.